cull deprecated nixos-hardware modules
May need the flake again someday in the future, but for now my hardware
is not in need of an upstream maintainer to manage. Thanks though, it
was really insightful to port the configs back over. Also, unless I
misunderstand, the settings I applied for now may update soon as these
deprecated modules get upstreamed.