fix lazyvim news unread state
This should be fixed upstream by moving the unread state of news entries
into ~/.local/state/nvim/lazy (or nearby). I think folke outright
rejected this idea in the following issue, but they didn't actually
answer the question.

https://github.com/LazyVim/LazyVim/issues/4149

Either way, I'm right, this is not configuration but state, and should
be treated as such instead of living in `lazyvim.json` which also should
be considered state, not configuration, because what it does is maintain
state based on selections in `:Lazy`. Run-on sentences be damned. Fix
this shit.