reduce profanity sureness
"yank" was being flagged, but is a common word for clipboard/register
operations. maybe this should just be a 1-line exclusion directive
comment instead. this is my personal config, and generally won't be
using any words in a profane context anyway, so we'll just use a weaker
filter for now.