remove gamemode
gamemoderun within steam is a headache, have to add it for every single
game's launch options. instead, we launch steam under gamemode and it
works for the games launched as a child of that.

alternatively, can edit the libraryconfig.vdf file, but this looks like
an awful hack and has to be redone every time a game is installed
anyway. it's not really better than just setting the launch option