add virtual displays
these don't seem to be activating yet, but this should allow me to run
special headless screens for sunshine.

another problem to solve is pushing steam to the correct display & tag.
river does not expose the necessary information easily to pull steam
over, so it may depend on killing and restarting steam in the right
place. this is very slow if steam was not already running, but checking
if the process is running is also non-trivial. procps package is not
available in sunshine's env somehow, or pgrep just fails for some other
reason