remove alex
this doesn't understand source code very well. we could use the --text
option but there is no effective way to disable directives on an line
basis, and profanity cannot be overridden at all. only sureness can be
evaluated, which is a completely mixed bag. like the project, impl needs
improvement